ipc: Call ipm_set_enabled() also for dual IPM support
The IPC drivers rpmsg_service and rpmsg_multi_instance are not
explicitly enabling the IPM channels when two different devices are used
for TX and RX. While this could be redundant for some IPM drivers, in
some cases the hardware needs to be enabled before using it.

Add the missing calls to ipm_set_enabled() for both the devices.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>