ipc: icmsg: add a function to send a message from multiple fragments
Sometimes a message to send is scattered in the memory. The caller of
the sending function may gather the message to a continuous buffer by
copying the fragments to a buffer. But such approach is inefficient
because of unnecessary copy operation and sending buffer allocation.

This patch adds a capability for the icmsg library to send a message
consisting of fragments scattered in the memory, accessing the fragments
directly.

Signed-off-by: Hubert Miś <hubert.mis@gmail.com>