ipc: icmsg multi endpoint implementation optimization
The implementation of multi endpoint icmsg (one of ipc_service
backends) performed copying for the message to send as a scatter-
gather operation. With support to send fragmented message by the
icmsg library it is no longer required to copy the message or to
allocate a buffer to send.

This patch is an optimization for the icmsg-me backend improving
memory efficiency by removing send buffer from each instance of the
backend and improving run-time performance by skipping one copying
of the message to send.

Signed-off-by: Hubert Miś <hubert.mis@gmail.com>